LPS-71559 Declare the dependency and do not allow wrappers to be active if their store is null
