+++ modules/apps/foundation/portal-store/portal-store-safe-file-name-wrapper/src/main/java/com/liferay/portal/store/safe/file/name/wrapper/SafeFileNameAdvancedFileSystemStoreWrapper.java
@@ -12,29 +12,38 @@
  * details.
  */
 
-package com.liferay.document.library.internal.store;
+package com.liferay.portal.store.safe.file.name.wrapper;
 
 import com.liferay.document.library.kernel.store.Store;
 import com.liferay.document.library.kernel.store.StoreWrapper;
+import com.liferay.portal.store.safe.file.name.SafeFileNameStore;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Adolfo Pérez
+ * @author Roberto Díaz
  */
 @Component(
 	immediate = true,
 	property = {
-		"store.type=com.liferay.portal.store.file.system.AdvancedFileSystemStore",
-		"store.type=com.liferay.portal.store.file.system.FileSystemStore"
+		"store.type=com.liferay.portal.store.file.system.AdvancedFileSystemStore"
 	},
 	service = StoreWrapper.class
 )
-public class IgnoreDuplicatesStoreWrapper implements StoreWrapper {
+public class SafeFileNameAdvancedFileSystemStoreWrapper
+	implements StoreWrapper {
+
+	@Reference(
+		target = "(store.type=com.liferay.portal.store.file.system.AdvancedFileSystemStore)",
+		unbind = "-"
+	)
+	public void setStore(Store store) {
+	}
 
 	@Override
 	public Store wrap(Store store) {
-		return new IgnoreDuplicatesStore(store);
+		return new SafeFileNameStore(store);
 	}
 
 }
